created dto for transaction entity

+++ wallet-service/src/main/java/com/nerdysoft/walletservice/dto/request/TransactionRequestDto.java
@@ -1,5 +1,6 @@
 package com.nerdysoft.walletservice.dto.request;
 
 import com.nerdysoft.walletservice.model.enums.Currency;
+import java.math.BigDecimal;
 
-public record TransactionRequestDto(Double amount, Currency currency) {}
+public record TransactionRequestDto(BigDecimal amount, Currency currency) {}
