+++ wallet-service/src/main/java/com/nerdysoft/walletservice/model/Transaction.java
@@ -9,6 +9,7 @@ import jakarta.persistence.Enumerated;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 import lombok.Getter;
@@ -28,7 +29,7 @@ public class Transaction {
   private UUID toWalletId;
 
   @Column(nullable = false)
-  private Double amount;
+  private BigDecimal amount;
 
   @Enumerated(EnumType.STRING)
   private Currency currency;
@@ -38,14 +39,14 @@ public class Transaction {
 
   private final LocalDate createdAt = LocalDate.now();
 
-  public Transaction(UUID walletId, Double amount, Currency currency, TransactionStatus status) {
+  public Transaction(UUID walletId, BigDecimal amount, Currency currency, TransactionStatus status) {
     this.walletId = walletId;
     this.amount = amount;
     this.currency = currency;
     this.status = status;
   }
 
-  public Transaction(UUID walletId, Double amount, Currency currency, TransactionStatus status, UUID toWalletId) {
+  public Transaction(UUID walletId, BigDecimal amount, Currency currency, TransactionStatus status, UUID toWalletId) {
     this(walletId, amount, currency, status);
     this.toWalletId = toWalletId;
   }
